requirements.txt
Flask
request
bs4